removed the uint64 cast.

+++ src/runtime/proc.go
@@ -1931,7 +1931,7 @@ func oneNewExtraM() {
 		// while calling from C thread to Go.
 		traceGoCreate(gp, 0) // no start pc
 		gp.traceseq++
-		traceEvent(traceEvGoInSyscall, -1, uint64(gp.goid))
+		traceEvent(traceEvGoInSyscall, -1, gp.goid)
 	}
 	// put on allg for garbage collector
 	allgadd(gp)
